Extract independent native MCP administration endpoints
