+++ packages/agent-native-bridge/agent-native-bridge.cabal
@@ -78,6 +78,7 @@ foreign-library haskell-agent-bridge
                     , Agent.CLI.MacOS.NativeRequest
                     , Agent.CLI.MacOS.NativeGatewayBoundary
                     , Agent.CLI.MacOS.NativeModelCatalog
+                    , Agent.CLI.MacOS.McpAdminBridge
                     , Agent.CLI.MacOS.RepositoryInput
                     , Agent.CLI.MacOS.ResourceAdmin
     c-sources:        cbits/HaskellAgentBridgeRuntime.c
@@ -165,6 +166,7 @@ test-suite agent-native-bridge-test
                      , Agent.CLI.MacOS.NativeRequest
                      , Agent.CLI.MacOS.NativeGatewayBoundary
                      , Agent.CLI.MacOS.NativeModelCatalog
+                     , Agent.CLI.MacOS.McpAdminBridge
                     , Agent.CLI.MacOS.ResourceAdmin
                      , Agent.CLI.MacOS.BrowserBridgeFFISpec
                      , Agent.CLI.MacOS.ComputerBridgeSpec
